# swagger.yaml
openapi: 3.0.3
info:
  title: Dodo Payments Backend Engine API
  version: 1.0.0
  description: High-performance, multi-tenant billing and payment aggregation engine.
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Environment
paths:
...
  /invoices/{id}/pay:
    post:
      summary: Attempt Payment Execution
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            example: "req_uniq_prod_01"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - card_token
              properties:
                card_token:
                  type: string
                  example: "tok_visa_success"
      responses:
        "200":
          description: Payment successful, invoice transitioned to PAID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  invoice_id:
                    type: string
                    format: uuid
                  psp_reference:
                    type: string
                    format: uuid
                example:
                  status: "success"
                  invoice_id: "c8e1219b-6b80-4ad3-9e4c-1d1136b81ef1"
                  psp_reference: "42a39281-b541-4712-9ee1-88fc2139ba80"
        "202":
          description: Gateway timeout encountered. Processing asynchronously via polling.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "processing"
                  message:
                    type: string
                    example: "Gateway timeout reached. State is checking asynchronously."
        "422":
          description: Operational conflict (e.g., row-lock, duplicate charge attempt, or card declined).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - type
        - code
        - message
        - timestamp
      properties:
        type:
          type: string
          description: High-level categorization classification.
          example: "TRANSACTION_ERROR"
        code:
          type: string
          description: Specific programmatic error identifier string.
          example: "CARD_DECLINED"
        message:
          type: string
          description: Clear human-readable problem explanation statement.
          example: "The bank gateway rejected the charge due to insufficient funds."
        timestamp:
          type: string
          format: date-time
          example: "2026-05-24T03:56:15Z"
...